Remove repeated logic to traits.
