+++ includes/Dashboard.php
@@ -28,12 +28,15 @@
 
 namespace Google\Web_Stories;
 
+use Google\Web_Stories\TRAITS\Assets;
 use WP_Screen;
 
 /**
  * Dashboard class.
  */
 class Dashboard {
+
+	use Assets;
 	/**
 	 * Script handle.
 	 *
@@ -145,20 +148,7 @@ class Dashboard {
 			return;
 		}
 
-		$asset_file   = WEBSTORIES_PLUGIN_DIR_PATH . 'assets/js/' . self::SCRIPT_HANDLE . '.asset.php';
-		$asset        = is_readable( $asset_file ) ? require $asset_file : [];
-		$dependencies = isset( $asset['dependencies'] ) ? $asset['dependencies'] : [];
-		$version      = isset( $asset['version'] ) ? $asset['version'] : WEBSTORIES_VERSION;
-
-		wp_enqueue_script(
-			self::SCRIPT_HANDLE,
-			WEBSTORIES_PLUGIN_DIR_URL . 'assets/js/' . self::SCRIPT_HANDLE . '.js',
-			$dependencies,
-			$version,
-			false
-		);
-
-		wp_set_script_translations( self::SCRIPT_HANDLE, 'web-stories' );
+		$this->load_assert( self::SCRIPT_HANDLE );
 
 		$rest_base     = Story_Post_Type::POST_TYPE_SLUG;
 		$new_story_url = admin_url(
@@ -258,7 +248,7 @@ class Dashboard {
 			self::SCRIPT_HANDLE,
 			WEBSTORIES_PLUGIN_DIR_URL . 'assets/css/' . self::SCRIPT_HANDLE . '.css',
 			[ 'google-fonts' ],
-			$version
+			WEBSTORIES_VERSION
 		);
 
 		// Dequeue forms.css, see https://github.com/google/web-stories-wp/issues/349 .
