+++ includes/Embed_Block.php
@@ -26,10 +26,14 @@
 
 namespace Google\Web_Stories;
 
+use Google\Web_Stories\TRAITS\Assets;
+
 /**
  * Embed block class.
  */
 class Embed_Block {
+
+	use Assets;
 	/**
 	 * Script handle.
 	 *
@@ -53,26 +57,13 @@ class Embed_Block {
 		wp_register_script( 'amp-story-player', 'https://cdn.ampproject.org/amp-story-player-v0.js', [], 'v0', false );
 		wp_register_style( 'amp-story-player', 'https://cdn.ampproject.org/amp-story-player-v0.css', [], 'v0' );
 
-		$asset_file   = WEBSTORIES_PLUGIN_DIR_PATH . 'assets/js/' . self::SCRIPT_HANDLE . '.asset.php';
-		$asset        = is_readable( $asset_file ) ? require $asset_file : [];
-		$dependencies = isset( $asset['dependencies'] ) ? $asset['dependencies'] : [];
-		$version      = isset( $asset['version'] ) ? $asset['version'] : WEBSTORIES_VERSION;
-
-		$dependencies[] = 'amp-story-player';
-
-		wp_register_script(
-			self::SCRIPT_HANDLE,
-			WEBSTORIES_PLUGIN_DIR_URL . 'assets/js/' . self::SCRIPT_HANDLE . '.js',
-			$dependencies,
-			$version,
-			false
-		);
+		$this->load_assert( self::SCRIPT_HANDLE, [ 'amp-story-player' ] );
 
 		wp_register_style(
 			self::STYLE_HANDLE,
 			WEBSTORIES_PLUGIN_DIR_URL . 'assets/css/' . self::STYLE_HANDLE . '.css',
 			[ 'amp-story-player' ],
-			$version
+			WEBSTORIES_VERSION
 		);
 
 		// todo: use register_block_type_from_metadata() once generally available.
